Corrección int>float
Usando tipo de dato int/long (Norm o unsigned), se rompía a partir del n°13 en adelante

+++ G2E9.cpp
@@ -6,7 +6,7 @@ using namespace std;
 
 int main(int argc, char** argv) {
 
-    int n, r;
+    float n, r;
     r = 1;
     cout << "Ingrese el valor de n" << endl;
     cin >> n;
